Conflicts:
	Source/EventFlow/Core/ITransientFaultHandler.cs
	Source/EventFlow/ReadStores/InMemory/InMemoryReadStore.cs
	Source/EventFlow/ReadStores/MultipleAggregateReadStoreManager.cs
	Source/EventFlow/ReadStores/ReadStoreManager.cs
	Source/EventFlow/ReadStores/SingleAggregateReadStoreManager.cs

+++ Source/EventFlow.EventStores.EventStore/EventStoreEventPersistence.cs
@@ -1,8 +1,8 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015-2016 Rasmus Mikkelsen
-// Copyright (c) 2015-2016 eBay Software Foundation
-// https://github.com/rasmus/EventFlow
+// Copyright (c) 2015-2018 Rasmus Mikkelsen
+// Copyright (c) 2015-2018 eBay Software Foundation
+// https://github.com/eventflow/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
@@ -20,7 +20,7 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// 
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -72,7 +72,8 @@ namespace EventFlow.EventStores.EventStore
                 resolvedEvents.AddRange(allEventsSlice.Events.Where(e => !e.OriginalStreamId.StartsWith("$")));
                 nextPosition = allEventsSlice.NextPosition;
 
-            } while (resolvedEvents.Count < pageSize && !allEventsSlice.IsEndOfStream);
+            }
+            while (resolvedEvents.Count < pageSize && !allEventsSlice.IsEndOfStream);
 
             var eventStoreEvents = Map(resolvedEvents);
 
@@ -171,10 +172,11 @@ namespace EventFlow.EventStores.EventStore
                     200,
                     false)
                     .ConfigureAwait(false);
-                nextSliceStart = currentSlice.NextEventNumber;
+                nextSliceStart = (int)currentSlice.NextEventNumber;
                 streamEvents.AddRange(currentSlice.Events);
 
-            } while (!currentSlice.IsEndOfStream);
+            }
+            while (!currentSlice.IsEndOfStream);
 
             return Map(streamEvents);
         }
@@ -189,7 +191,7 @@ namespace EventFlow.EventStores.EventStore
             return resolvedEvents
                 .Select(e => new EventStoreEvent
                     {
-                        AggregateSequenceNumber = e.Event.EventNumber + 1, // Starts from zero
+                        AggregateSequenceNumber = (int)(e.Event.EventNumber + 1), // Starts from zero
                         Metadata = Encoding.UTF8.GetString(e.Event.Metadata),
                         AggregateId = e.OriginalStreamId,
                         Data = Encoding.UTF8.GetString(e.Event.Data),
